searcher/requirements.txt
fastapi==0.128.0
uvicorn[standard]==0.40.0
python-multipart==0.0.21
pydantic==2.12.5
pydantic-settings==2.12.0

haystack-ai==2.21.0
pgvector-haystack==5.4.0